Bug 870171 - Part 6: Move BroadcastReceiver to base FHR package; remove UploadReceiver.
