+++ manifests/HealthReportAndroidManifest_services.xml.in
@@ -6,7 +6,7 @@
           -->
         <service
             android:exported="false"
-            android:name="org.mozilla.gecko.background.healthreport.upload.HealthReportBroadcastService" >
+            android:name="org.mozilla.gecko.background.healthreport.HealthReportBroadcastService" >
         </service>
         <service
             android:exported="false"
