+++ src/main/java/org/mozilla/gecko/background/healthreport/HealthReportBroadcastReceiver.java
@@ -2,10 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.gecko.background.healthreport.upload;
+package org.mozilla.gecko.background.healthreport;
 
 import org.mozilla.gecko.background.common.log.Logger;
-import org.mozilla.gecko.background.healthreport.HealthReportConstants;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
